==> 01 Get Data.ipynb <==
# 01 Get Data

## Imports and setup

%matplotlib inline

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt

from bs4 import BeautifulSoup
import json, pickle, random, re, requests
from typing import List, Dict, Any

from sentence_transformers import SentenceTransformer
from sklearn.metrics.pairwise import cosine_similarity
from umap import UMAP

SEED = 2023
random.seed(SEED)

## Final functions

# Secondary functions for the individual preprocessing steps, no need to run standalone

def scrape_links(url: str) -> List[str]:
    """
    Extracts all hyperlinks from a single url.
    """
    response = requests.get(url)
    soup = BeautifulSoup(response.text, "html.parser")
    links = [a_tag.get('href') for a_tag in soup.find_all('a')]

    return links


def scrape_links_from_multiple_urls(urls: List[str]) -> pd.DataFrame:
    """
    Extract all links from multiple urls and return pd.DataFrame.
    """
    df = pd.DataFrame(columns=['link', 'source_url'])
    for url in urls:
        new_rows = pd.DataFrame({'link': scrape_links(url), 'source_url': url})
        df = pd.concat([df, new_rows], ignore_index=True)
        
    return df


def clean_links(df: pd.DataFrame) -> pd.DataFrame:
    PREFIXES = [
        'Category:',
        'File:',
        'Help:',
        'Special:',
        'Talk:',
        'Wikipedia:',
        'Main_Page|Portal:',
        'List_of',
    ]
    LINKS_TO_REMOVE_PATTERN = fr'^/wiki/(?!{"|".join(PREFIXES)}).*'

    return (df
            .drop_duplicates()
            .dropna()
            # Filter links
            .loc[lambda df_: df_['link'].str.contains(LINKS_TO_REMOVE_PATTERN)]
            # Add `title` column 
            .assign(title=lambda df_: (df_['link']
                                       .str.replace('/wiki/', '')
                                       .str.replace('_', ' ')
                                       .str.capitalize()))
            .reset_index(drop=True))


def deep_dictionary_search(my_dict: Dict[Any, Any], my_key: Any) -> Any:
    """
    Utility function that recursively searches for a key at any level and returns its value.
    """
    for key, value in my_dict.items():
        if key == my_key:
            return value
        elif isinstance(value, dict):
            result = deep_dictionary_search(value, my_key)
            if result is not None:
                return result
    return None


def retrieve_summary(title: str) -> str:
    """
    Retrieves a summary of a Wikipedia article based on a given article title.
    """
    api_request_url = f'https://en.wikipedia.org/w/api.php?action=query&prop=extracts&titles={title}&format=json'
    response = requests.get(api_request_url)
    try:
        json_data = response.json()
        summary = deep_dictionary_search(json_data['query'], 'extract')
    except json.JSONDecodeError:
        summary = ''

    return summary


def retrieve_and_clean_summaries(df: pd.DataFrame, debug=True) -> pd.DataFrame:
    """
    Download and preprocess the article summaries.
    Caution! Long execution time!
    
    Args:
    `debug` - retrieve summaries for only 10 articles
    
    """
    HTML_CONTENT_PATTERN = re.compile('<.*?>')
    
    # 10 articles in case debug=True
    df = df.iloc[:10] if debug else df
    
    return (df
            # Retrieve summaries
            .assign(summary=lambda df_: df_['title'].apply(retrieve_summary))

            # Drop improperlly-retrieved summaries - they start with identical starting characters.
            .loc[lambda df_: ~df_['summary'].str[:20].duplicated()]

            # Strip HTML and white spaces
            .assign(summary=lambda df_: (df_['summary']
                                       .str.replace(HTML_CONTENT_PATTERN, '')
                                       .str.strip()))

            .reset_index(drop=True)
            )

# Main Function, combining the secondary functions from above.
# Run to scrape article links, retrive and clean article summaries.

def retrieve_manufacturing_articles(debug=False) -> pd.DataFrame:
    """
    Retrieves summaries of Wikipedia articles related to manufacturing.
    Relevant articles are obtained by scraping list-type articles with references.
    
    Parameters:
        debug (bool): If True, retrieves a small subset of articles for debugging purposes.
    
    Returns:
        pd.DataFrame: DataFrame containing the retrieved articles and their summaries.
    """
    
    # List of URLs for list-type articles related to manufacturing
    article_urls = [
        'https://en.wikipedia.org/wiki/List_of_welding_processes',
        'https://en.wikipedia.org/wiki/List_of_manufacturing_processes',
        'https://en.wikipedia.org/wiki/Industrial_processes',
    ]
    
    print('Scraping article links...')
    scraped_links = scrape_links_from_multiple_urls(article_urls)
    print('Completed!')
    
    # Filtering the scraped links to leave only the relevant ones
    print('Filtering relevant links...')
    cleaned_links = clean_links(scraped_links)
    print('Completed!')
        
    # Retrieve and clean the summaries of the articles
    print('Retrieving article summaries...')
    retrieved_articles = retrieve_and_clean_summaries(cleaned_links, debug=debug)
    print('Completed!')
    
    return retrieved_articles


# Test
retrieved_articles = retrieve_manufacturing_articles(debug=True)
retrieved_articles

# Full donwload
retrieved_articles = retrieve_manufacturing_articles(debug=False)

# Optionally save as csv
retrieved_articles.to_csv('data/manufacturing_articles.csv', index=False)

## Developing the algorithm

def scrape_links(url: str) -> List[str]:
    """
    Extracts all hyperlinks from a single url.
    """
    response = requests.get(url)
    soup = BeautifulSoup(response.text, "html.parser")
    links = [a_tag.get('href') for a_tag in soup.find_all('a')]

    return links


# Test
test_url = 'https://en.wikipedia.org/wiki/List_of_manufacturing_processes'

print(f'Starting test scraping from {test_url}...')
links = scrape_links(test_url)
print(f'Extracted total of {len(links)} links.')

def scrape_links_from_multiple_urls(urls: List[str]) -> pd.DataFrame:
    """
    Extract all links from multiple urls and return pd.DataFrame.
    """
    df = pd.DataFrame(columns=['link', 'source_url'])
    for url in urls:
        new_rows = pd.DataFrame({'link': scrape_links(url), 'source_url': url})
        df = pd.concat([df, new_rows], ignore_index=True)
        
    return df


urls = [
    'https://en.wikipedia.org/wiki/List_of_welding_processes',
    'https://en.wikipedia.org/wiki/List_of_manufacturing_processes',
    'https://en.wikipedia.org/wiki/Industrial_processes',
]

print(f'Starting scraping from {len(urls)} urls...')
all_links = scrape_links_from_multiple_urls(urls)
print(f'Extracted the following:')
all_links.groupby('source_url').count()

def clean_links(df: pd.DataFrame) -> pd.DataFrame:
    PREFIXES = [
        'Category:',
        'File:',
        'Help:',
        'Special:',
        'Talk:',
        'Wikipedia:',
        'Main_Page|Portal:',
        'List_of',
    ]
    LINKS_TO_REMOVE_PATTERN = fr'^/wiki/(?!{"|".join(PREFIXES)}).*'

    return (df
            .drop_duplicates()
            .dropna()
            # Filter links
            .loc[lambda df_: df_['link'].str.contains(LINKS_TO_REMOVE_PATTERN)]
            # Add `title` column 
            .assign(title=lambda df_: (df_['link']
                                       .str.replace('/wiki/', '')
                                       .str.replace('_', ' ')
                                       .str.capitalize()))
            .reset_index(drop=True))


print(f'Starting cleaning...')

cleaned_links = clean_links(all_links)

print(f'Clean complete! Remaining links:')
cleaned_links.groupby('source_url').count()

# Optional: store the DataFrame
# df.to_csv('data/manufacturing_processes_links.csv', index=False)

def deep_dictionary_search(my_dict: Dict[Any, Any], my_key: Any) -> Any:
    """
    Utility function that recursively searches for a key at any level and returns its value.
    """
    for key, value in my_dict.items():
        if key == my_key:
            return value
        elif isinstance(value, dict):
            result = deep_dictionary_search(value, my_key)
            if result is not None:
                return result
    return None


def retrieve_summary(title: str) -> str:
    """
    Retrieves a summary of a Wikipedia article based on a given article title.
    """
    api_request_url = f'https://en.wikipedia.org/w/api.php?action=query&prop=extracts&titles={title}&format=json'
    response = requests.get(api_request_url)
    try:
        json_data = response.json()
        summary = deep_dictionary_search(json_data['query'], 'extract')
    except json.JSONDecodeError:
        summary = ''

    return summary


def retrieve_and_clean_summaries(df: pd.DataFrame, debug=True) -> pd.DataFrame:
    """
    Download and preprocess the article summaries.
    Caution! Long execution time!
    
    Args:
    `debug` - retrieve summaries for only 10 articles
    
    """
    HTML_CONTENT_PATTERN = re.compile('<.*?>')
    
    # 10 articles in case debug=True
    df = df.iloc[:10] if debug else df
    
    return (df
            # Retrieve summaries
            .assign(summary=lambda df_: df_['title'].apply(retrieve_summary))

            # Drop improperlly-retrieved summaries - they start with identical starting characters.
            .loc[lambda df_: ~df_['summary'].str[:20].duplicated()]

            # Strip HTML and white spaces
            .assign(summary=lambda df_: (df_['summary']
                                       .str.replace(HTML_CONTENT_PATTERN, '')
                                       .str.strip()))

            .reset_index(drop=True)
            )
    
# Test
print(f'Starting retrieving article summaries from Wikipedia...')

links_with_summaries = retrieve_and_clean_summaries(cleaned_links)

print(f'Done! Sample articles and their summaries:')
pd.options.display.max_colwidth = 400
display(links_with_summaries.loc[:, ['title', 'summary']].sample(2))
pd.options.display.max_colwidth = 50
