==> 3 Reduce Dimensions.ipynb <==
# 03 Dimensionality Reduction

## Imports and setup

%matplotlib inline

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt

import pickle, random

import bokeh, random

from umap import UMAP

SEED = 2023
random.seed(SEED)

## Load embeddings and data

with open('data/transformer_embeddings.pickle', 'rb') as f:
    embeddings = pickle.load(f)
    
print(f'Loaded article embeddings! Shape: {embeddings.shape}')

df = pd.read_csv('data/manufacturing_articles.csv')
print(f'Loaded pd.DataFrame with the article titles and their summaries.')
print(f'Shape: {df.shape}')
df.head(2)

## Reduce dimensions using UMAP

Uniform Manifold Approximation and Projection

my_umap = UMAP(n_components=2)
umap_embeddings = my_umap.fit_transform(embeddings)

## Export data for `bulk`

embeddings_2D = pd.DataFrame({
    'text': df['title'],
    'x': umap_embeddings[:, 0],
    'y': umap_embeddings[:, 1],
})

embeddings_2D.head()

# Save
embeddings_2D.to_csv("embeddings/umap_2d.csv", index=False)

## Plot using `bokeh`

Alternative `bokeh` plot in this notebook in case `bulk` is not used.

# Load
embeddings_2D = pd.read_csv("embeddings/umap_2d.csv")

from bokeh.plotting import figure, show
from bokeh.models import HoverTool, ColumnDataSource, WheelZoomTool
from bokeh.io import output_notebook

# Enable Bokeh to display plots in the JupyterLab notebook
output_notebook()

# Data
data = embeddings_2D

# Create a ColumnDataSource from the data dictionary
source = ColumnDataSource(data)

# Create the figure
p = figure(tools='hover, box_zoom', tooltips=[('Text', '@text')])

# Plot the points
p.circle(x='x', y='y', source=source, size=10)

# Add hover tool
hover = p.select(dict(type=HoverTool))
hover.tooltips = [
    ('Text', '@text'),
    ('X', '@x'),
    ('Y', '@y')
]

# Show the plot
show(p)
