==> 4 Annotate.ipynb <==
# 4 Annotate
_Annotation (label assignment) three ways: 1) topic modeling, 2) manual, 3) few-shot learining._

## Imports and setup

%matplotlib inline

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt

from bertopic import BERTopic
from sklearn.feature_extraction.text import CountVectorizer

documents = pd.read_csv('data/manufacturing_articles.csv')
documents

embeddings = pd.read_csv('embeddings/umap_2d.csv')
embeddings

embeddings[['x', 'y']].to_numpy()

vectorizer_model = CountVectorizer(stop_words='english')
topic_model = BERTopic(vectorizer_model=vectorizer_model)

topics, probs = topic_model.fit_transform(
    documents['summary'],
    embeddings[['x', 'y']].to_numpy()
)
